fix(online-scoring): address review on judge-answer diagnostics
- Keep the judge's raw answer out of the rule's logs, which are persisted per
  workspace: NOT_JSON / NOT_A_JSON_OBJECT now report the answer's size, and the
  answer itself stays in the server log. Matches "Never Log" in
  .agents/skills/opik-backend/SKILL.md and AgenticScoringService.summarizeResponse,
  which reports shape rather than text for the same reason. NO_SCORE_FIELDS keeps
  reporting the field names: they are structural and are the diagnostic that
  matters.
- Cap the reported field names. They come from the judge's answer, so one reply
  with thousands of top-level fields would otherwise write a huge row.
- Copy the lists in toParsed() so a caller cannot mutate a parsed result through
  the accumulator's own lists.
- Structure the internal logs as key='value', dropping the static "expected
  structure" text that repeated on every occurrence and carried no per-event
  information. The user-facing messages stay prose: they render on the rule's
  Logs page for a customer, not in ops telemetry.

Move the parsing tests to OnlineScoringEngineParsingTest. toFeedbackScores and
logUnreadableResponse are pure static functions, but OnlineScoringEngineTest
starts MySQL, ClickHouse and Zookeeper for its end-to-end tests, so asserting
that a quoted score parses required a Docker daemon and ~45s. The 30 parsing
tests now run in under a second; the end-to-end Redis-flow test stays behind.
